requirements.txt
# 项目依赖（固定版本，避免兼容性问题）
requests==2.31.0
pyyaml==6.0.1
pytest==7.4.0  # 测试依赖